centralized wildcardToRegex and doubleSlashes methods
